Fix #3265: setting default index to 0
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@33932 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/operators/OnionPeel/OnionPeelAttributes.java
@@ -80,7 +80,7 @@ public class OnionPeelAttributes extends AttributeSubject implements Plugin
         categoryName = new String("Whole");
         subsetName = new String("Whole");
         index = new Vector();
-        index.addElement(new Integer(1));
+        index.addElement(new Integer(0));
         logical = false;
         requestedLayer = 0;
         seedType = SEEDIDTYPE_SEEDCELL;
@@ -96,7 +96,7 @@ public class OnionPeelAttributes extends AttributeSubject implements Plugin
         categoryName = new String("Whole");
         subsetName = new String("Whole");
         index = new Vector();
-        index.addElement(new Integer(1));
+        index.addElement(new Integer(0));
         logical = false;
         requestedLayer = 0;
         seedType = SEEDIDTYPE_SEEDCELL;
